+++ app/src/App/index.tsx
@@ -21,7 +21,7 @@ import { ProtocolRunDetails } from '../pages/Devices/ProtocolRunDetails'
 import { RobotSettings } from '../pages/Devices/RobotSettings'
 import { ProtocolsLanding } from '../pages/Protocols/ProtocolsLanding'
 import { ProtocolDetails } from '../pages/Protocols/ProtocolDetails'
-import { AppSettings } from '../organisms/AppSettings'
+import { AppSettings } from '../pages/AppSettings'
 import { Labware } from '../pages/Labware'
 import { Navbar } from './Navbar'
 import { PortalRoot as ModalPortalRoot, TopPortalRoot } from './portal'
